* changed pk + fk in db to int(10)

+++ config/db/tables.sql
@@ -3,12 +3,12 @@
 --
 
 CREATE TABLE `ext_search_filtercondition` (
-  `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
+  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id_person_create` int(10) unsigned NOT NULL DEFAULT '0',
   `deleted` tinyint(2) NOT NULL DEFAULT '0',
-  `id_set` smallint(5) unsigned NOT NULL,
+  `id_set` int(10) unsigned NOT NULL,
   `filter` varchar(64) NOT NULL,
   `value` varchar(100) NOT NULL,
   `negate` tinyint(1) NOT NULL DEFAULT '0',
@@ -22,12 +22,12 @@ CREATE TABLE `ext_search_filtercondition` (
 --
 
 CREATE TABLE `ext_search_filterset` (
-  `id` smallint(5) unsigned NOT NULL AUTO_INCREMENT,
+  `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id_person_create` int(10) unsigned NOT NULL DEFAULT '0',
   `deleted` tinyint(2) NOT NULL DEFAULT '0',
-  `sorting` smallint(5) unsigned NOT NULL,
+  `sorting` int(10) unsigned NOT NULL,
   `is_hidden` tinyint(2) NOT NULL DEFAULT '0',
   `roles` varchar(16) NOT NULL,
   `type` varchar(16) NOT NULL,
